feat: modify deploy scripts
