+++ contracts/alliance-hub/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "alliance-hub"
-version = "0.1.0"
+version = "0.1.1"
 authors = ["Terra Money <core@terra.money>"]
 edition = "2018"
