+++ scripts/package.json
@@ -1,6 +1,6 @@
 {
   "dependencies": {
-    "@terra-money/feather.js": "1.0.7-beta.4",
+    "@terra-money/feather.js": "2.0.0-beta.7",
     "dotenv": "^16.1.4",
     "ts-node": "^10.9.1"
   }
